+++ contracts/netmap/migration_test.go
@@ -18,7 +18,7 @@ import (
 const name = "netmap"
 
 func TestMigration(t *testing.T) {
-	err := dump.IterateDumps("../testdata", func(id dump.ID, r *dump.Reader) {
+	err := dump.IterateDumps("../../testdata", func(id dump.ID, r *dump.Reader) {
 		t.Run(id.String()+"/"+name, func(t *testing.T) {
 			testMigrationFromDump(t, r)
 		})
@@ -58,9 +58,6 @@ func testMigrationFromDump(t *testing.T, d *dump.Reader) {
 		},
 	})
 
-	require.NotZerof(t, balanceHash, "missing storage item %q with Balance contract address", balanceHashOldKey)
-	require.NotZerof(t, containerHash, "missing storage item %q with Container contract address", containerHashOldKey)
-
 	updPrm := []any{
 		false,
 		util.Uint160{}, // Balance contract
@@ -96,8 +93,8 @@ func testMigrationFromDump(t *testing.T, d *dump.Reader) {
 	prevPendingVote := readPendingVotes()
 
 	// read previous values using contract API
-	readUint64 := func(method string) uint64 {
-		n, err := c.Call(t, method).TryInteger()
+	readUint64 := func(method string, args ...any) uint64 {
+		n, err := c.Call(t, method, args...).TryInteger()
 		require.NoError(t, err)
 		return n.Uint64()
 	}
@@ -184,7 +181,11 @@ func testMigrationFromDump(t *testing.T, d *dump.Reader) {
 
 	c.CheckUpdateSuccess(t, updPrm...)
 
-	checkNewEpochSubscribers(t, c, balanceHash, containerHash)
+	if prevVersion < 19_000 {
+		require.NotZerof(t, balanceHash, "missing storage item %q with Balance contract address", balanceHashOldKey)
+		require.NotZerof(t, containerHash, "missing storage item %q with Container contract address", containerHashOldKey)
+		checkNewEpochSubscribers(t, c, balanceHash, containerHash)
+	}
 
 	// check that contract was updates as expected
 	newPendingVotes := readPendingVotes()
@@ -200,11 +201,16 @@ func testMigrationFromDump(t *testing.T, d *dump.Reader) {
 	require.Nil(t, c.GetStorageItem(notaryDisabledKey), "notary flag should be removed")
 	require.Nil(t, c.GetStorageItem([]byte("innerring")), "Inner Ring nodes should be removed")
 	require.Equal(t, prevCurrentEpoch, newCurrentEpoch, "current epoch should remain")
-	require.Equal(t, prevCurrentEpochBlock, newCurrentEpochBlock, "current epoch block should remain")
+	require.Equal(t, prevCurrentEpochBlock, newCurrentEpochBlock, "current epoch block should remain (method)")
+	require.Nil(t, c.GetStorageItem([]byte("snapshotBlock")), "current epoch block should be removed (storage)")
+	require.EqualValues(t, prevCurrentEpochBlock, readUint64("getEpochBlock", prevCurrentEpoch),
+		"current epoch block should be resolvable")
 	require.ElementsMatch(t, prevConfigs, newConfigs, "config should remain")
 	require.ElementsMatch(t, prevCurrentNetmap, newCurrentNetmap, "current netmap should remain")
 	require.ElementsMatch(t, prevNetmapCandidates, newNetmapCandidates, "netmap candidates should remain")
 	require.ElementsMatch(t, ir, c.InnerRing(t))
+	require.Nil(t, c.GetStorageItem(balanceHashOldKey), "balance contract address should be removed")
+	require.Nil(t, c.GetStorageItem(containerHashOldKey), "container contract address should be removed")
 
 	require.Equal(t, len(prevDiffToSnapshots), len(newDiffToSnapshots))
 	for k, vPrev := range prevDiffToSnapshots {
@@ -215,13 +221,10 @@ func testMigrationFromDump(t *testing.T, d *dump.Reader) {
 
 	var cleanupThreshItem = c.GetStorageItem([]byte("t"))
 	cleanupThresh := io.NewBinReaderFromBuf(cleanupThreshItem).ReadVarUint()
-	require.Equal(t, 3, cleanupThresh)
+	require.EqualValues(t, 3, cleanupThresh)
 }
 
 func checkNewEpochSubscribers(t *testing.T, contract *migration.Contract, balanceWant, containerWant util.Uint160) {
-	require.Nil(t, contract.GetStorageItem(balanceHashOldKey))
-	require.Nil(t, contract.GetStorageItem(containerHashOldKey))
-
 	// contracts are migrated in alphabetical order at least for now
 
 	var balanceMigrated bool